reverseInPlace and reversed changed
The condition in for loop in reverseInPlace() has been changed to divide the array by 2 and a temp variable is added to save the old variable and replaces the reversed index.

The for loop in reversed is changed so the index is in decreasing order for the array list to print values located at the end of the index first then move up.